Define an error for a failed (return status != 200) request to the IDP.
If an incorrect password or other reason causes a non 200 response
from the IDP, the response may not be valid XML. If the IDP's ECP
endpoint is protected by Apache, it may send an HTML response.
Previously in that case, the error returned by login would be an XML
parser error. However, the root cause of the error is that the
request to the IDP was not successful. See #5.